refactor: cleanup filter fieldtype constants
