+++ desk/src/components/List/ListFilterValue.vue
@@ -34,6 +34,13 @@
 </template>
 
 <script setup>
+import {
+	numberTypes,
+	dateTypes,
+	timespanOptionsList,
+	checkOptionsList,
+	isOptionsList,
+} from "@/stores/list_filter"
 import { ref, watch } from "vue"
 import { FormControl, DatePicker } from "frappe-ui"
 
@@ -67,107 +74,10 @@ const emit = defineEmits(["update"])
 watch(value, (value) => {
 	emit("update", value)
 })
-
-const timespanOptionsList = [
-	{
-		label: "Last Week",
-		value: "last week",
-	},
-	{
-		label: "Last Month",
-		value: "last month",
-	},
-	{
-		label: "Last Quarter",
-		value: "last quarter",
-	},
-	{
-		label: "Last 6 Months",
-		value: "last 6 months",
-	},
-	{
-		label: "Last Year",
-		value: "last year",
-	},
-	{
-		label: "Yesterday",
-		value: "yesterday",
-	},
-	{
-		label: "Today",
-		value: "today",
-	},
-	{
-		label: "Tomorrow",
-		value: "tomorrow",
-	},
-	{
-		label: "This Week",
-		value: "this week",
-	},
-	{
-		label: "This Month",
-		value: "this month",
-	},
-	{
-		label: "This Quarter",
-		value: "this quarter",
-	},
-	{
-		label: "This Year",
-		value: "this year",
-	},
-	{
-		label: "Next Week",
-		value: "next week",
-	},
-	{
-		label: "Next Month",
-		value: "next month",
-	},
-	{
-		label: "Next Quarter",
-		value: "next quarter",
-	},
-	{
-		label: "Next 6 Months",
-		value: "next 6 months",
-	},
-	{
-		label: "Next Year",
-		value: "next year",
-	},
-]
-
-const checkOptionsList = [
-	{
-		label: "Yes",
-		value: "true",
-	},
-	{
-		label: "No",
-		value: "false",
-	},
-]
-
-const isOptionsList = [
-	{
-		label: "Set",
-		value: "set",
-	},
-	{
-		label: "Not Set",
-		value: "not set",
-	},
-]
-
 const optionsList = {
 	timespan: timespanOptionsList,
 	is: isOptionsList,
 	Check: checkOptionsList,
 	Select: props.options,
 }
-
-const numberTypes = ["Float", "Int", "Currency", "Percent"]
-const dateTypes = ["Date", "Datetime"]
 </script>
